* Task/Bug Id: Import PVs without JBoss restart
* Code Reviewed By:  Pavan K 		On:  19th Jan 2010

* Technical Details:  Updated cache

* Unit tested(Y/N):  Yes

* Unit Test Case(s) Id/Name: TestImportPermissibleValue

* Developer Notes: Added code for cache update and added feature of bulk PVs upload

* Current Build++ Counts for this file : No changes in PMD count
* PMD - X     Checkstyle - X     Findbugs - X     CPD – X

SVN-Revision: 26271

+++ software/DynamicExtentions/src/java/edu/common/dynamicextensions/util/DirOperationsUtility.java
@@ -35,7 +35,8 @@ public final class DirOperationsUtility
 		return dirOperations;
 	}
 
-	private DirOperationsUtility(){
+	private DirOperationsUtility()
+	{
 
 	}
 
@@ -47,14 +48,21 @@ public final class DirOperationsUtility
 	 */
 	public void createNewTempDirectory(String tempDirName) throws DynamicExtensionsSystemException
 	{
-		File tempDir = new File(tempDirName);
-		if (tempDir.exists())
+		if ((tempDirName.length()!=0) && tempDirName != null)
 		{
-			deleteDirectory(new File(tempDirName));
+			File tempDir = new File(tempDirName);
+			if (tempDir.exists())
+			{
+				deleteDirectory(new File(tempDirName));
+			}
+			if (!tempDir.mkdirs())
+			{
+				throw new DynamicExtensionsSystemException("Unable to create tempDirectory");
+			}
 		}
-		if (!tempDir.mkdirs())
+		else
 		{
-			throw new DynamicExtensionsSystemException("Unable to create tempDirectory");
+			LOGGER.info("Given temp directory is empty");
 		}
 	}
